Added basic layout and dynamic background image functionality
Created HTML structure with a top section for "Crypto" and "Weather", a main time display, and an author credit area.

Applied CSS styling for layout, including full-height main container, flexbox alignment for top section, and responsive background size.

Implemented a `fetch` request to Unsplash's API to get a random nature photos and set them  as the background image of the page’s body. Updated the `#author` element with the photo author's name and
added error handling to use a default fallback image in case of a failed fetch request.